// src/app/auth/login/schemas.ts
import * as z from 'zod'

export const LoginSchema = z.object({
  email: z.string().email({
    message: 'Please enter a valid email address'
  }),
  password: z.string().min(1, {
    message: 'Please enter a password'
  }),
  code: z.optional(z.string())
})

export type LoginValues = z.infer<typeof LoginSchema>